--- pages/api/v1/ticket/transfer.js
@@ -1,31 +1,0 @@
-const { prisma } = require("../../../../prisma/prisma");
-
-export default async function transferTicket(req,res) {
-    console.log(req.body);
-  try {
-    await prisma.ticket.update({
-      where: { id: Number(req.body.find) },
-      data: {
-        assignedTo: {
-          connect: { id: Number(req.body.id) },
-        },
-      },
-    });
-
-    const tickets = await prisma.ticket.findMany({
-      where: { isIssued: true, userId: Number(req.user.id) },
-      include: {
-        client: {
-          select: { id: true, name: true },
-        },
-        assignedTo: {
-          select: { id: true, firstName: true, lastName: true },
-        },
-      },
-    });
-    res.status(200).json({ tickets });
-  } catch (error) {
-    console.log(error);
-    return res.status(500);
-  }
-}
